ai-data-integration/requirements.txt
Django==5.0.1
django-cors-headers==4.2.0
django-storages==1.14.4
djangorestframework==3.14.0
python-dotenv
drf_spectacular
psycopg2-binary
fitz 
pydantic
openai
instructor
gunicorn 
boto3
PyMuPDF==1.24.13